split null transform into null and loopback
